Remove file_info hashes, corpus_metadata
They really aren't used, and conflict with the vision of providing a
simple CHAT -> YAML transcription service. This will hopefully rectify
it.

Similarly, @ages have been removed as well, since there isn't any
indication we'll be using that specific functionality (but it could
change if Tim wants it implemented again).