IMPORTANT: Remove verb-specific functionality
This is an important commit, which gets rid of a lot of the code
required to create the verbs by ages file that this script was used
previously. It follows the current development's trajectory of making
this a script limited strictly to YAML -> CHAT conversion.

See the verb-functionality branch if this behavior is required again.